+++ src/LocalDateTime.php
@@ -94,16 +94,6 @@ final class LocalDateTime implements JsonSerializable
 
     /**
      * Creates a LocalDateTime from a native DateTime or DateTimeImmutable object.
-     *
-     * @deprecated please use fromNativeDateTime instead
-     */
-    public static function fromDateTime(DateTimeInterface $dateTime): LocalDateTime
-    {
-        return self::fromNativeDateTime($dateTime);
-    }
-
-    /**
-     * Creates a LocalDateTime from a native DateTime or DateTimeImmutable object.
      */
     public static function fromNativeDateTime(DateTimeInterface $dateTime): LocalDateTime
     {
@@ -692,40 +682,10 @@ final class LocalDateTime implements JsonSerializable
      *
      * Note that the native DateTime object supports a precision up to the microsecond,
      * so the nanoseconds are rounded down to the nearest microsecond.
-     *
-     * @deprecated please use toNativeDateTime instead
-     */
-    public function toDateTime(): DateTime
-    {
-        return $this->toNativeDateTime();
-    }
-
-    /**
-     * Converts this LocalDateTime to a native DateTime object.
-     *
-     * The result is a DateTime in the UTC time-zone.
-     *
-     * Note that the native DateTime object supports a precision up to the microsecond,
-     * so the nanoseconds are rounded down to the nearest microsecond.
      */
     public function toNativeDateTime(): DateTime
     {
         return $this->atTimeZone(TimeZone::utc())->toNativeDateTime();
-    }
-
-    /**
-     * Converts this LocalDateTime to a native DateTimeImmutable object.
-     *
-     * The result is a DateTimeImmutable in the UTC time-zone.
-     *
-     * Note that the native DateTimeImmutable object supports a precision up to the microsecond,
-     * so the nanoseconds are rounded down to the nearest microsecond.
-     *
-     * @deprecated please use toNativeDateTimeImmutable instead
-     */
-    public function toDateTimeImmutable(): DateTimeImmutable
-    {
-        return $this->toNativeDateTimeImmutable();
     }
 
     /**