+++ src/TimeZoneRegion.php
@@ -124,10 +124,7 @@ final class TimeZoneRegion extends TimeZone
         return $this->zone->getOffset($dateTime);
     }
 
-    /**
-     * @deprecated please use toNativeDateTimeZone instead
-     */
-    public function toDateTimeZone(): DateTimeZone
+    public function toNativeDateTimeZone(): DateTimeZone
     {
         return clone $this->zone;
     }